pyproject.toml
[project]
name = "django-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "aiofiles>=24.1.0",
  "aiohttp>=3.13.5",
  "aiojobs>=1.4.0",
  "argon2-cffi>=25.1.0",
  "celery>=5.5.3",
  "django>=6.0.4",
  "django-csp>=4.0",
  "django-extra-checks>=0.16.1",
  "django-http-referrer-policy>=1.1.1",
  "django-migration-linter>=5.2.0",
  "django-ninja>=1.4.3",
  "django-permissions-policy>=4.28.0",
  "django-redis>=6.0.0",
  "django-split-settings>=1.3.2",
  "django-stubs-ext>=5.2.2",
  "django-test-migrations>=1.5.0",
  "psycopg[binary]>=3.2.10",
  "python-decouple>=3.8",
  "rq>=2.6.0",
  "structlog>=25.4.0",
  "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
  "django-debug-toolbar>=6.3.0",
  "django-querycount>=0.8.3",
  "django-stubs>=6.0.2",
  "ipython>=9.5.0",
  "mypy>=1.17.1",
  "nplusone>=1.0.0",
  "pytest>=9.0.3",
  "pytest-django>=4.11.1",
  "pytest-randomly>=4.0.0",
  "ruff>=0.13.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
  "A",    # flake8-builtins
  "ARG",  # flake8-unused-arguments
  "ASYNC",
  "B",    # flake8-bugbear
  "BLE",  # flake8-blind-except
  "C4",   # flake8-comprehensions
  "C90",  # mccabe
  "DTZ",  # flake8-datetimez
  "E",    # pycodestyle errors
  "EM",   # flake8-errmsg
  "ERA",  # eradicate
  "F",    # pyflakes
  "FA",   # flake8-future-annotations
  "FBT",  # flake8-boolean-trap
  "I",    # isort
  "INP",  # flake8-no-pep420
  "ISC",  # flake8-implicit-str-concat
  "LOG",  # flake8-logging
  "N",    # pep8-naming
  "PERF", # perflint
  "PIE",  # flake8-pie
  "PL",   # pylint
  "PTH",  # flake8-use-pathlib
  "PT",   # flake8-pytest-style
  "Q",    # flake8-quotes
  "RET",  # flake8-return
  "RUF",  # ruff
  "S",    # flake8-bandit
  "SIM",  # flake8-simplify
  "T20",  # flake8-print
  "TCH",  # flake8-type-checking
  "TID",  # flake8-tidy-imports
  "TRY",  # tryceratops
  "UP",   # pyupgrade
  "W",    # pycodestyle warnings
]

ignore = [
  "Q000", # Quote style handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"*/settings/*" = [
  "ERA001",
  "FBT003",
  "PLW1508",
  "S104", # bind all interfaces in dev examples
  "S105", # settings use config() not raw secrets
]
"tests/*" = [
  "ARG001",
  "ARG002",
  "INP001",
  "PLR0913",
  "PLR2004",
  "S101", # assert in tests
  "S105",
  "S106",
]
"*/migrations/*" = [
  "E501",
  "ERA001",
  "I001",
  "INP001",
  "RUF012",
  "S",
]

[tool.ruff.lint.mccabe]
max-complexity = 8

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
show_error_codes = true
files = ["apps", "server", "manage.py", "tests"]
mypy_path = ["."]
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = "decouple"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "server.settings"
